Handle quoting of / with \; properly handle __ in paths, quote " and \n in
Mon Mar 19 15:53:36 2001  Owen Taylor  <otaylor@redhat.com>

	* gtk/gtkitemfactory.c gtk/testgtk.c: Handle quoting of / with \; properly
	handle __ in paths, quote " and \n in gtk_item_factory_dump_items().

	* gtk/gtkadjustment.[ch] gtk/gtkclist.c gtk/gtkhruler.c gtk/gtklayout.c
	  gtk/gtklist.c gtk/gtkprogress.[ch] gtk/gtkprogressbar.[ch] gtk/gtkrange.[ch]
	  gtk/gtkruler.[ch] gtk/gtkscale.c gtk/gtkscrolledwindow.c gtk/gtkspinbutton.[ch]
	  gtk/gtktext.c gtk/gtktextview.c gtk/gtkvruler.c gtk/testgtk.c:
	Change float to double everywhere with the exception of 0<->1.0 alignment
	and GtkCurve.

+++ gtk/gtkadjustment.h
@@ -52,12 +52,12 @@ struct _GtkAdjustment
 {
   GtkObject parent_instance;
   
-  gfloat lower;
-  gfloat upper;
-  gfloat value;
-  gfloat step_increment;
-  gfloat page_increment;
-  gfloat page_size;
+  gdouble lower;
+  gdouble upper;
+  gdouble value;
+  gdouble step_increment;
+  gdouble page_increment;
+  gdouble page_size;
 };
 
 struct _GtkAdjustmentClass
@@ -70,19 +70,19 @@ struct _GtkAdjustmentClass
 
 
 GtkType	   gtk_adjustment_get_type		(void) G_GNUC_CONST;
-GtkObject* gtk_adjustment_new			(gfloat		  value,
-						 gfloat		  lower,
-						 gfloat		  upper,
-						 gfloat		  step_increment,
-						 gfloat		  page_increment,
-						 gfloat		  page_size);
+GtkObject* gtk_adjustment_new			(gdouble	  value,
+						 gdouble	  lower,
+						 gdouble	  upper,
+						 gdouble	  step_increment,
+						 gdouble	  page_increment,
+						 gdouble	  page_size);
 void	   gtk_adjustment_changed		(GtkAdjustment	 *adjustment);
 void	   gtk_adjustment_value_changed		(GtkAdjustment	 *adjustment);
 void	   gtk_adjustment_clamp_page		(GtkAdjustment	 *adjustment,
-						 gfloat		  lower,
-						 gfloat		  upper);
+						 gdouble	  lower,
+						 gdouble	  upper);
 void	   gtk_adjustment_set_value		(GtkAdjustment	 *adjustment,
-						 gfloat		  value);
+						 gdouble	  value);
 
 #ifdef __cplusplus
 }
